+++ raspberry_lights_controller/lib/service/pattern.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 import 'package:dio/dio.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:raspberry_lights_controller/models/pattern_configuration.dart';
+import 'package:raspberry_lights_controller/models/pattern_configuration_setting.dart';
 import 'package:raspberry_lights_controller/providers/current_pattern.dart';
 import 'package:raspberry_lights_controller/providers/network.dart';
 import 'package:raspberry_lights_controller/providers/pattern_list.dart';
@@ -29,14 +30,7 @@ Future<void> setLightPattern(
     'brightness': patternConfiguration.brightness,
     'additionalSettings': selectedPattern.additionalSettings.isNotEmpty
         ? patternConfiguration.additionalSettings
-              .map(
-                (setting) => PatternConfigurationSetting(
-                  name: setting.name,
-                  value: setting.value,
-                ),
-              )
-              .toList()
-        : <PatternConfiguration>[],
+        : <PatternConfigurationSetting>[],
   };
 
   await client.post<void>(